# dsa-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the API for the DSA Backend application.",
        "title": "DSA Backend API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/problem/crud/create": {
            "put": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Create a new lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Create a new lecture entry",
                "parameters": [
                    {
                        "description": "Lecture entry details",
                        "name": "lectureEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/problem.LectureEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry created successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/create/{lectureid}/{problemid}": {
            "post": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Register a new problem associated with a lecture",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Register a new problem",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Problem ID",
                        "name": "problemid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Zip file contains problem resources",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problem registered successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/delete/{lectureid}": {
            "delete": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Delete an existing lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Delete an existing lecture entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/delete/{lectureid}/{problemid}": {
            "delete": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "delete a problem entry by lecture ID and problem ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "delete problem entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Problem ID",
                        "name": "problemid",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Problem deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "404": {
                        "description": "Problem not found",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
        "/problem/crud/update/{lectureid}": {
            "patch": {
                "security": [
                    {
                        "OAuth2Password": [
                            "grading"
                        ]
                    }
                ],
                "description": "Update an existing lecture entry, accessible by manager and admin.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update an existing lecture entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Lecture ID",
                        "name": "lectureid",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Lecture entry details",
                        "name": "lectureEntry",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/problem.LectureEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lecture entry updated successfully",
                        "schema": {
                            "$ref": "#/definitions/response.Success"
                        }
                    },
                    "400": {
                        "description": "Invalid request",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "problem.LectureEntryRequest": {
            "type": "object",
            "required": [
                "deadline",
                "id",
                "start_date",
                "title"
            ],
            "properties": {
                "deadline": {
                    "type": "string",
                    "default": "2025-12-01T10:00:00+09:00"
                },
                "id": {
                    "type": "integer",
                    "default": 0
                },
                "start_date": {
                    "type": "string",
                    "default": "2025-10-01T10:00:00+09:00"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "response.Error": {
            "type": "object",
            "properties": {
                "errors": {
                    "type": "object",
                    "properties": {
                        "body": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "response.Success": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "OAuth2Password": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/api/user/login",
            "scopes": {
                "admin": "Grants any rights related to admin user",
                "manager": "Grants any rights related to manager user",
                "me": "Grants any rights related to the current user"
            }
        }
    }
}